requirements.txt
# Core FastAPI dependencies
groq>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0

# OpenAI API client
openai>=1.0.0

# Google Generative AI
google-generativeai>=0.3.0

# AI and ML dependencies
openai-whisper>=20231117
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0

# Audio processing
pyaudio>=0.2.11
SpeechRecognition>=3.8.1
pydub>=0.25.1
gtts>=2.3.0

# Medical data processing
scikit-learn>=1.3.0
scipy>=1.10.0
matplotlib>=3.7.0

# Database and storage - PostgreSQL support
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
redis>=4.5.0

# Communication and notifications
requests>=2.31.0
aiohttp>=3.8.0
celery>=5.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Logging and monitoring
loguru>=0.7.0
prometheus-client>=0.17.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
